media_service/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
Pillow==12.1.0
aiofiles==25.1.0
requests==2.32.3
